+++ app/Http/Controllers/TournamentController.php
@@ -29,9 +29,9 @@ class TournamentController extends Controller
             'participation_fee' => 'nullable|numeric|min:0',
             'phase' => 'nullable|string|max:50',
             'description' => 'nullable|string',
-            'logo' => 'nullable|image|mimes:jpeg,png,svg,jpg|max:2048',
-            'banner' => 'nullable|image|mimes:jpeg,png,svg,jpg|max:5120',
-            'stream_banner' => 'nullable|image|mimes:jpeg,svg,png,jpg|max:5120'
+            'logo' => 'nullable|image|mimes:webp,jpeg,png,svg,jpg|max:2048|ratio:1,1',
+            'banner' => 'nullable|image|mimes:webp,jpeg,png,svg,jpg|max:5120|ratio:3,1',
+            'stream_banner' => 'nullable|image|mimes:webp,jpeg,svg,png,jpg|max:5120|ratio:16,9',
         ];
     }
     public function index()
@@ -42,6 +42,7 @@ class TournamentController extends Controller
 
     public function create(Club $club)
     {
+        $this->authorize('create_tournament', $club);
         $layout = request()->header('X-Ajax-Request') ? 'layouts.ajax' : 'layouts.app';
 
         $nextFriday = Carbon::now()->next(Carbon::FRIDAY);
@@ -58,27 +59,21 @@ class TournamentController extends Controller
             'layout' => $layout,
             'mode' => 'create',
             'club' => $club,
+            'styles' => ['tournament-form.css'],
             'tournament' => $tournament,
             'tournamentPhases' => Tournament::PHASES,
         ]);
     }
 
-    public function store(Request $request)
+    public function store(Request $request, Club $club)
     {
-//        $validated = $request->validate([
-//            'club_id' => 'sometimes|exists:clubs,id',
-//            'name' => 'required|string|max:255',
-//            'date_start' => 'required|date',
-//            'date_end' => 'required|date|after_or_equal:date_start',
-//            'logo' => 'nullable|image',
-//            'banner' => 'nullable|image',
-//            'stream_banner' => 'nullable|image'
-//        ]);
+        $this->authorize('create_tournament', $club);
 
         $validated = $request->validate($this->rules());
 
-        if (!isset($validated['club_id']) && $request->route('club')) {
-            $validated['club_id'] = $request->route('club');
+//        if (!isset($validated['club_id']) && $request->route('club')) {
+        if (!isset($validated['club_id'])) {
+            $validated['club_id'] = $club->id;
         }
 
         $files = ['logo', 'banner', 'stream_banner'];
@@ -87,7 +82,8 @@ class TournamentController extends Controller
                 $validated[$file] = $request->file($file)->store("tournament-{$file}s", 'public');
             }
         }
-        $validated['duration'] = $validated['date_start']->diffInDays($validated['date_end']) + 1;
+        $validated['duration'] = \Carbon\Carbon::parse($validated['date_start'])
+                ->diffInDays(\Carbon\Carbon::parse($validated['date_end'])) + 1;
 
         $tournament = Tournament::create($validated);
 
@@ -99,12 +95,15 @@ class TournamentController extends Controller
 
     public function show(Club $club, Tournament $tournament)
     {
+        $this->authorize('manage_tournament', $tournament);
         $layout = request()->header('X-Ajax-Request') ? 'layouts.ajax' : 'layouts.app';
 
         return view('tournaments.form', [
             compact('tournament'),
             'layout' => $layout,
             'mode' => 'show',
+
+            'styles' => ['tournament-form.css'],
             'club' => $club,
             'tournament' => $tournament,
             'tournamentPhases' => Tournament::PHASES,
@@ -113,12 +112,14 @@ class TournamentController extends Controller
 
     public function edit(Club $club, Tournament $tournament)
     {
+        $this->authorize('manage_tournament', $tournament);
         $layout = request()->header('X-Ajax-Request') ? 'layouts.ajax' : 'layouts.app';
 
         return view('tournaments.form', [
             compact('tournament'),
             'layout' => $layout,
             'tournament' => $tournament,
+            'styles' => ['tournament-form.css'],
             'mode' => 'edit',
             'club' => $club,
             'tournamentPhases' => Tournament::PHASES,
@@ -128,6 +129,18 @@ class TournamentController extends Controller
 
     public function update(Request $request, Club $club, Tournament $tournament)
     {
+//        dd($request->all());
+        $this->authorize('manage_tournament', $tournament);
+//        if ($request->hasFile('logo')) {
+//            $file = $request->file('logo');
+//            dd([
+//                'file_name' => $file->getClientOriginalName(),
+//                'extension' => $file->getClientOriginalExtension(),
+//                'mime_type' => $file->getMimeType(),
+//                'size' => $file->getSize(),
+//                'is_valid' => $file->isValid()
+//            ]);
+//        }
 
         $validated = $request->validate($this->rules());
 
@@ -154,8 +167,9 @@ class TournamentController extends Controller
             ->with('tab', 'tournaments');
     }
 
-    public function destroy(Tournament $tournament)
+    public function destroy(Club $club, Tournament $tournament)
     {
+        $this->authorize('manage_tournament', $tournament);
         $tournament->delete();
         return redirect()->route('tournaments.index')->with('success', 'Tournament deleted successfully.');
     }
